Fix default internal telemetry endpoint for collector
    https://github.com/open-telemetry/opentelemetry-operator/issues/3730

+++ pkg/collector/upgrade/v0_111_0_test.go
@@ -76,7 +76,7 @@ func Test0_111_0Upgrade(t *testing.T) {
 										"exporter": map[string]any{
 											"prometheus": map[string]any{
 												"host": "0.0.0.0",
-												"port": 8888,
+												"port": int32(8888),
 											},
 										},
 									},
